requirements.txt
openai>=1.0.0
requests>=2.31.0
playwright>=1.40.0
pyyaml>=6.0
schedule>=1.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
aiohttp>=3.9.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
